requirements.txt
aiogram
google-generativeai==0.8.3
aiohttp
Pillow